backend-flask/pyproject.toml
[project]
name = "backend-flask"
version = "0.1.0"
description = ""
authors = [
    {name = "Chan, Wei Jian Ivan",email = "ivan_chan@mymail.sutd.edu.sg"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "mistralai (>=1.8.1,<2.0.0)",
    "aws-wsgi (>=0.2.7,<0.3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "backend_flask", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
